requirements.txt
asgiref==3.7.2
Django==5.0.1
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
drf-nested-routers==0.93.5
mysqlclient==2.2.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0

djoser~=2.2.2